// packages/user-connectrpc-server/src/generated/chat3_user_pb.d.ts
export declare type DialogMember = Message$1<"chat3.user.DialogMember"> & {
  /**
   * @generated from field: string user_id = 1;
   */
  userId: string;

  /**
   * @generated from field: google.protobuf.Struct meta = 2;
   */
  meta?: JsonObject;

  /**
   * @generated from field: chat3.user.MemberState state = 3;
   */
  state?: MemberState;
};

/**
 * Describes the message chat3.user.DialogMember.
 * Use `create(DialogMemberSchema)` to create a new message.
 */
export declare const DialogMemberSchema: GenMessage<DialogMember>;

/**
 * Состояние участника
 *
 * @generated from message chat3.user.MemberState
 */
export declare type MemberState = Message$1<"chat3.user.MemberState"> & {
  /**
   * @generated from field: int32 unread_count = 1;
   */
  unreadCount: number;

  /**
   * @generated from field: double last_seen_at = 2;
   */
  lastSeenAt: number;

  /**
   * @generated from field: double last_message_at = 3;
   */
  lastMessageAt: number;

  /**
   * @generated from field: bool is_active = 4;
   */
  isActive: boolean;
};

/**
 * Describes the message chat3.user.MemberState.
 * Use `create(MemberStateSchema)` to create a new message.
 */
export declare const MemberStateSchema: GenMessage<MemberState>;

/**
 * Сообщение
 *
 * @generated from message chat3.user.Message
 */
export declare type Message = Message$1<"chat3.user.Message"> & {
  /**
   * @generated from field: string message_id = 1;
   */
  messageId: string;

  /**
   * @generated from field: string dialog_id = 2;
   */
  dialogId: string;

  /**
   * @generated from field: string sender_id = 3;
   */
  senderId: string;

  /**
   * @generated from field: string type = 4;
   */
  type: string;

  /**
   * @generated from field: string content = 5;
   */
  content: string;

  /**
   * @generated from field: google.protobuf.Struct meta = 6;
   */
  meta?: JsonObject;

  /**
   * @generated from field: repeated chat3.user.MessageStatus statuses = 7;
   */
  statuses: MessageStatus[];

  /**
   * @generated from field: google.protobuf.Struct reaction_set = 8;
   */
  reactionSet?: JsonObject;

  /**
   * @generated from field: chat3.user.SenderInfo sender_info = 9;
   */
  senderInfo?: SenderInfo;

  /**
   * @generated from field: double created_at = 10;
   */
  createdAt: number;

  /**
   * @generated from field: string topic_id = 11;
   */
  topicId: string;

  /**
   * @generated from field: google.protobuf.Struct topic = 12;
   */
  topic?: JsonObject;
};

/**
 * Describes the message chat3.user.Message.
 * Use `create(MessageSchema)` to create a new message.
 */
export declare const MessageSchema: GenMessage<Message>;

/**
 * Статус сообщения
 *
 * @generated from message chat3.user.MessageStatus
 */
export declare type MessageStatus = Message$1<"chat3.user.MessageStatus"> & {
  /**
   * @generated from field: string user_id = 1;
   */
  userId: string;

  /**
   * @generated from field: string status = 2;
   */
  status: string;

  /**
   * @generated from field: double read_at = 3;
   */
  readAt: number;

  /**
   * @generated from field: double created_at = 4;
   */
  createdAt: number;
};

/**
 * Describes the message chat3.user.MessageStatus.
 * Use `create(MessageStatusSchema)` to create a new message.
 */
export declare const MessageStatusSchema: GenMessage<MessageStatus>;

/**
 * Информация об отправителе
 *
 * @generated from message chat3.user.SenderInfo
 */
export declare type SenderInfo = Message$1<"chat3.user.SenderInfo"> & {
  /**
   * @generated from field: string user_id = 1;
   */
  userId: string;

  /**
   * @generated from field: string name = 2;
   */
  name: string;

  /**
   * @generated from field: double created_at = 3;
   */
  createdAt: number;

  /**
   * @generated from field: google.protobuf.Struct meta = 4;
   */
  meta?: JsonObject;
};